Added translations to tooltips of sort and minimize/maximize

+++ ActionNote.Common/Converters/ImportantWidthConverter.cs
@@ -5,8 +5,8 @@ namespace ActionNote.Common.Converters
 {
     public class ImportantWidthConverter : IValueConverter
     {
-        private const int IMPORTANT_WIDTH = 12;
-        private const int UNIMPORTANT_WIDTH = 8;
+        private const double IMPORTANT_WIDTH = 12;
+        private const double UNIMPORTANT_WIDTH = 8;
 
         public object Convert(object value, Type targetType, object parameter, string language)
         {
